Revert any changes not related to refactor
